Updated ecommerce.js, kindly consider
This comes from the documentation:
http://api.jquery.com/on/
Delegated event handlers have the advantage that they can process events from descendant elements that are added to the document at a later time. By picking an element that is guaranteed to be present at the time the delegated event handler is attached, you can use delegated events to avoid the need to frequently attach and remove event handlers. This element could be the container element of a view in a Model-View-Controller design, for example, or document if the event handler wants to monitor all bubbling events in the document. The document element is available in the head of the document before loading any other HTML, so it is safe to attach events there without waiting for the document to be ready.

And this comes from stack overflow (credits: Popnoodles):
https://stackoverflow.com/questions/33522627/jquery-rebind-click-event-after-ajax-call
"
$(staticAncestors).on(eventName, dynamicChild, function() {});

For example, if your page was dynamically creating elements with the class name dosomething you would bind the event to a parent which already exists (this is the nub of the problem here, you need something that exists to bind to, don't bind to the dynamic content), this can be (and the easiest option) is document. Though bear in mind document may not be the most efficient option."

In my opinion using document isn't a bad option even performance wise.